render-backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
pdf2docx
aiofiles
python-docx
PyMuPDF
Pillow
reportlab
docx2txt
